[Rename] Fetch 파일 삭제 이후 해당되는 코드 Dashboard 파일에 추가
